external-toolchain.bbclass: stop overriding S, we need it for debugsrc
We need to be able to copy sources from the external toolchain into it in
do_unpack, then back out into the install area in do_install.

Note that attempts were made to bypass this indirection by having do_package
directly copy sources from the external toolchain, and while this did function,
it results in pseudo failures breaking the build. Further investigation could be
done on this in the future.

JIRA: SB-21854

Signed-off-by: Christopher Larson <chris_larson@mentor.com>