external-debugsrc.bbclass: add class
This class copies debug sources out of the external toolchain, obeying the paths
in DEBUG_PREFIX_MAP, the same way do_package does to copy those sources into the
src binary package. Between the two, this redistributes the debug sources in the
appropriate -src packages as it would for non-external recipes.

JIRA: SB-21854

Signed-off-by: Christopher Larson <chris_larson@mentor.com>